move AsyncFileReader to separate mod
