+++ analytic_engine/src/sst/factory.rs
@@ -2,86 +2,23 @@
 
 //! Factory for different kinds sst builder and reader.
 
-use std::{
-    fmt::Debug,
-    ops::Range,
-    sync::{Arc, RwLock},
-};
+use std::{fmt::Debug, sync::Arc};
 
-use async_trait::async_trait;
-use bytes::Bytes;
 use common_types::projected_schema::ProjectedSchema;
-use common_util::{
-    error::{GenericError, GenericResult},
-    runtime::Runtime,
-};
+use common_util::runtime::Runtime;
 use object_store::{ObjectStoreRef, Path};
 use table_engine::predicate::PredicateRef;
 
 use crate::{
     sst::{
         builder::SstBuilder,
+        file_reader::FileReaderOnObjectStore,
         meta_data::cache::MetaCacheRef,
-        parquet::{
-            async_reader::AsyncFileChunkReader, builder::ParquetSstBuilder, AsyncParquetReader,
-            ThreadedReader,
-        },
+        parquet::{builder::ParquetSstBuilder, AsyncParquetReader, ThreadedReader},
         reader::SstReader,
     },
     table_options::{Compression, StorageFormat, StorageFormatHint},
 };
-
-pub struct FileReaderOnObjectStore {
-    path: Path,
-    store: ObjectStoreRef,
-    cached_file_size: RwLock<Option<usize>>,
-}
-
-impl FileReaderOnObjectStore {
-    pub fn new(path: Path, store: ObjectStoreRef) -> Self {
-        Self {
-            path,
-            store,
-            cached_file_size: RwLock::new(None),
-        }
-    }
-}
-
-#[async_trait]
-impl AsyncFileChunkReader for FileReaderOnObjectStore {
-    async fn file_size(&self) -> GenericResult<usize> {
-        // check cached filed_size first
-        {
-            let file_size = self.cached_file_size.read().unwrap();
-            if let Some(s) = file_size.as_ref() {
-                return Ok(*s);
-            }
-        }
-
-        // fetch the size from the underlying store
-        let head = self
-            .store
-            .head(&self.path)
-            .await
-            .map_err(|e| Box::new(e) as GenericError)?;
-        *self.cached_file_size.write().unwrap() = Some(head.size);
-        Ok(head.size)
-    }
-
-    async fn get_byte_range(&self, range: Range<usize>) -> GenericResult<Bytes> {
-        self.store
-            .get_range(&self.path, range)
-            .await
-            .map_err(|e| Box::new(e) as _)
-    }
-
-    async fn get_byte_ranges(&self, ranges: &[Range<usize>]) -> GenericResult<Vec<Bytes>> {
-        self.store
-            .get_ranges(&self.path, ranges)
-            .await
-            .map_err(|e| Box::new(e) as _)
-    }
-}
 
 /// Pick suitable object store for different scenes.
 pub trait ObjectStorePicker: Send + Sync + Debug {
