+++ analytic_engine/src/sst/mod.rs
@@ -5,6 +5,7 @@
 pub mod builder;
 pub mod factory;
 pub mod file;
+pub mod file_reader;
 pub mod manager;
 pub mod meta_data;
 pub mod metrics;
